Context managers for terminal setup/restore.
A context manager is a common idiom to define blocks with
guaranteed resource release. They can be defined as classes but
there's a nicer syntax using contextlib.contextmanager that lets
you define a function with a single `yield` in the middle, so you
can acquire resources before that and release them after.

Also after extracting those it seems that the thing with termios
attrs is not needed as you don't set anything before doing the work,
so there's probably nothing to restore. But I'm not the judge here.