Moving run() out of TetrisInputProcessor.
This is going to be the main theme of the refactoring. Since it's
not Java, there's no need to define everything as classes. It makes
Python code less readable with all those millions of `self.` and it
encourages usage of class attributes instead of simple local vars.
A rule of thumb is: if you're creating an single instance of a class
that lives through the whole program lifetime it shouldn't be a
class. Same goes for classes with a single public method.

Here's what's changed here:

- run() has been moved out of the class and lost its `self`
- instead run() creates an instance of the TetrisInputProcessor,
  which in turn will probably go away in the future as having a
  whole class to hold a single var looks like an overkill.